requirements.txt
aiogram>=3.7.0
vosk==0.3.45
pydantic-settings>=2.2.1
loguru>=0.7.2
google-genai>=1.0.0
aiohttp>=3.9.1
